% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanpumpRHelp-package.R
\docType{package}
\name{stanpumpRHelp-package}
\alias{stanpumpRHelp}
\alias{stanpumpRHelp-package}
\title{stanpumpRHelp: Examples and Help companion app for stanpumpR}
\description{
A lightweight Shiny application serving the Examples and Help content linked
from the main stanpumpR app. It is deployed separately to shinyapps.io and
reached from stanpumpR via its \code{help_link} configuration.
}
\author{
\strong{Maintainer}: Steven Shafer \email{steven.shafer@stanford.edu} (\href{https://orcid.org/0000-0002-0289-7793}{ORCID})

Authors:
\itemize{
  \item Steven Shafer \email{steven.shafer@stanford.edu} (\href{https://orcid.org/0000-0002-0289-7793}{ORCID})
}

}
\keyword{internal}
